Add support to open multiple shortlinks
